+++ src/Compute/Compute/ChangeLog.md
@@ -24,6 +24,10 @@
 * Updated Azure.Core from 1.45.0 to 1.47.3
 * Added `-EnableAutomaticUpgrade` and `-TreatFailureAsDeploymentFailure` parameters (Bool) to `New-AzVmGalleryApplication` and `New-AzVmssGalleryApplication` cmdlets.
 * Added `-EnableAutomaticUpgrade` and `-TreatFailureAsDeploymentFailure` parameters (Switch) to `Add-AzVmGalleryApplication` and `Add-AzVmssGalleryApplication` cmdlets.
+* Added `-ScheduledEventsApiVersion` and `-EnableAllInstancesDown` parameters to support Scheduled Events policy configuration
+    - Added to `Update-AzVmss`, `New-AzVmss`, `Update-AzVM`, `New-AzVM`, `Update-AzAvailabilitySet`, and `New-AzAvailabilitySet` cmdlets
+    - `ScheduledEventsApiVersion` parameter specifies the API (Application Programming Interface) version to determine which scheduled events schema version will be delivered
+    - `EnableAllInstancesDown` parameter specifies if scheduled events should be auto-approved when all Virtual Machine (VM) instances are down
 
 ## Version 10.5.0
 * Added `-Redeploy` switch parameter for `Update-AzHost` cmdlet to enable dedicated host redeployment.